requirements.txt
aiocache==0.11.1
discord==1.0.1
pipenv==2020.11.15
python-dotenv==0.17.0
PyYAML==5.4.1